ajustes + live tile

+++ IPOkemon/IPOkemon/Pokemon.cs
@@ -10,13 +10,13 @@ namespace IPOkemon
     {
         public string nombre { get; set; }
         public int nivel { get; set; }
-        public double exp { get; set; }
+        public int exp { get; set; }
         public string tipo { get; set; }
         public bool capturado { get; set; }
         public string descripcion { get; set; }
         public string sprite { get; set; }
 
-        public Pokemon(string nombre, int nivel, double exp, string tipo, bool capturado, string descripcion, string sprite)
+        public Pokemon(string nombre, int nivel, int exp, string tipo, bool capturado, string descripcion, string sprite)
         {
             this.nombre = nombre;
             this.nivel = nivel;
